resumeoptimiser/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# LLM / AI
openai==1.3.0

# Data Processing
numpy==1.26.3
pandas==2.1.3
scikit-learn==1.3.2

# Templates
jinja2==3.1.2

# LaTeX
pylatex==1.4.2

# HTTP Client
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Development (optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1